requirements.txt
langchain
langchain-chroma
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
python-dotenv
yfinance
newsapi-python
alpha-vantage
pandas
numpy